Restructured messages into separate files/modules.
