+++ server/src/actors/api/message/mod.rs
@@ -1,0 +1,9 @@
+mod connection_failed;
+mod error;
+mod global_chat;
+mod user_connected;
+
+pub use connection_failed::ConnectionFailed;
+pub use error::Error;
+pub use global_chat::GlobalChat;
+pub use user_connected::UserConnected;
