+++ server/src/actors/mod.rs
@@ -1,4 +1,7 @@
-pub mod api;
-pub mod game;
-pub mod player;
-pub mod server;
+mod api;
+mod game;
+mod player;
+mod server;
+
+pub use api::create_ws_api;
+pub use server::Server;
