+++ server/src/actors/player/message/connected.rs
@@ -1,0 +1,55 @@
+use actix::{Addr, AsyncContext, Context, Handler, Message};
+
+use super::super::super::api::{message as api_msg, ApiClient};
+use super::super::super::server::message as server_msg;
+use super::super::{Player, PlayerStatus};
+
+#[derive(Message)]
+#[rtype(result = "()")]
+pub struct Connected {
+    pub api_client: Addr<ApiClient>,
+}
+
+impl Handler<Connected> for Player {
+    type Result = ();
+
+    fn handle(&mut self, msg: Connected, ctx: &mut Context<Self>) -> Self::Result {
+        match self.status {
+            PlayerStatus::Connecting => {
+                self.status = PlayerStatus::Connected {
+                    api_client: msg.api_client.clone(),
+                };
+                msg.api_client.do_send(api_msg::UserConnected {
+                    session_id: self.session_id.clone(),
+                    username: self.username.clone(),
+                    player: ctx.address(),
+                });
+                self.server
+                    .do_send(server_msg::GlobalChatBroadcast::system_message(format!(
+                        "User '{}' connected.",
+                        self.username
+                    )));
+            }
+            PlayerStatus::LostConnection { .. } => {
+                self.status = PlayerStatus::Connected {
+                    api_client: msg.api_client.clone(),
+                };
+                msg.api_client.do_send(api_msg::UserConnected {
+                    session_id: self.session_id.clone(),
+                    username: self.username.clone(),
+                    player: ctx.address(),
+                });
+                self.server
+                    .do_send(server_msg::GlobalChatBroadcast::system_message(format!(
+                        "User '{}' reconnected.",
+                        self.username
+                    )));
+            }
+            PlayerStatus::Connected { .. } => {
+                msg.api_client.do_send(api_msg::ConnectionFailed {
+                    reason: "User still connected.".to_string(),
+                });
+            }
+        }
+    }
+}
