+++ server/src/actors/player/message/mod.rs
@@ -1,0 +1,9 @@
+mod connected;
+mod disconnected;
+mod receive_global_chat;
+mod send_global_chat;
+
+pub use connected::Connected;
+pub use disconnected::Disconnected;
+pub use receive_global_chat::ReceiveGlobalChat;
+pub use send_global_chat::SendGlobalChat;
