+++ server/src/actors/player/mod.rs
@@ -3,4 +3,4 @@ mod player;
 pub mod message;
 
 pub use player::Player;
-pub use player::PlayerStatus;
+use player::PlayerStatus;
